+++ src/seismometer/api/__init__.py
@@ -9,9 +9,11 @@ from .explore import *
 from .plots import *
 from .reports import *
 from .templates import *
+from .utils import *
 
 __all__ = ["ConfigProvider", "loader_factory", "Seismogram", "init_logger"]
 __all__.extend(explore.__all__)
-__all__.extend(reports.__all__)
 __all__.extend(plots.__all__)
+__all__.extend(reports.__all__)
 __all__.extend(templates.__all__)
+__all__.extend(utils.__all__)
